+++ nativeApp/src/screens/settings_views/Profile/index.tsx
@@ -1,7 +1,10 @@
-import React, { FC } from 'react';
+import React, { FC, useEffect } from 'react';
 
+import { useSelector, shallowEqual, useDispatch } from 'react-redux';
 import Page from '../../../lib/ui/Page';
 import UserForm from '../../../lib/ui/forms/UserForm';
+import { selectOrderedGenders, loadGenders } from '../../../redux/constants/genders';
+import { selectOrderedBirthYears } from '../../../redux/constants/birthYears';
 
 /*
  * Types
@@ -16,10 +19,25 @@ type Props = {
 /*
  * Component
  */
-const Profile: FC<Props> = () => (
-  <Page heading="Settings">
-    <UserForm onSubmit={() => {}} />
-  </Page>
-);
+const Profile: FC<Props> = () => {
+  const dispatch = useDispatch();
+  const genders = useSelector(selectOrderedGenders, shallowEqual);
+  const birthYears = useSelector(selectOrderedBirthYears, shallowEqual);
+
+  useEffect(() => {
+    if (!genders) {
+      dispatch(loadGenders());
+    }
+  });
+  return (
+    <Page heading="Settings">
+      <UserForm
+        onSubmit={() => {}}
+        genders={genders}
+        birthYears={birthYears}
+      />
+    </Page>
+  );
+};
 
 export default Profile;
